pyproject.toml
[tool.poetry]
name = "vulkan-platform-py"
version = "0.2.0"
description = ""
authors = ["Rayan Hatout <rayan.hatout@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10"
py-cpuinfo = "^8.0.0"
GPUtil = "^1.4.0"
pydantic = "^1.9.1"
mypy = "^0.960"

[tool.poetry.dev-dependencies]
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
plugins = [
  "pydantic.mypy"
]

follow_imports = "silent"
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true
disallow_untyped_defs = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true
